+++ odoo/local-src/delivery_rounds/models/instance.py
@@ -1,25 +1,9 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Jacques-Etienne Baudoux <je@bcim.be>
-#    Copyright 2016 BCIM sprl, Camptocamp
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2016-2017 Jacques-Etienne Baudoux (BCIM)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import math
+from datetime import datetime
 
 from odoo import api, fields, models
 from odoo.exceptions import Warning as UserError
@@ -34,9 +18,22 @@ def float2time(value):
     return '%d:%02d' % (hour, minute)
 
 
+def time2float(value):
+    return value.hour + value.minute / 60.0
+
+
+def time_now(record):
+    tz_name = record._context.get('tz') or record.env.user.tz
+    if not tz_name:
+        raise UserError(
+            "Please configure your timezone in your user preferences")
+    return time2float(fields.Datetime.context_timestamp(
+        record, datetime.now()))
+
+
 class RoundInstance(models.Model):
     _name = 'round.instance'
-    _order = 'date desc, time asc'
+    _order = 'date desc, time_picking_planned asc'
     _rec_name = 'complete_name'
 
     name = fields.Char(
@@ -50,16 +47,27 @@ class RoundInstance(models.Model):
         required=True,
         states={'done': [('readonly', True)]},
         default=fields.Date.context_today)
-    time = fields.Float(
-        'Planned Time',
+
+    time_picking_planned = fields.Float(
+        'Planned Picking Start Time',
         states={'done': [('readonly', True)]},
         )
-    vehicle_id = fields.Many2one(
-        'round.vehicle', 'Vehicle',
+    time_leave_planned = fields.Float(
+        'Planned Vehicle Start Time',
+        states={'done': [('readonly', True)]},
+        )
+
+    stat_time_picking = fields.Float(
+        'Picking Start Time', readonly=True)
+    stat_time_leave = fields.Float(
+        'Vehicle Start Time', readonly=True)
+
+    template_id = fields.Many2one(
+        'round.template', 'Template',
         states={'done': [('readonly', True)]},
         ondelete='restrict')
     color = fields.Integer(
-        related='vehicle_id.color')
+        related='template_id.color')
     state = fields.Selection(
         [('draft', 'Draft'),
          ('open', 'Confirmed'),
@@ -67,9 +75,9 @@ class RoundInstance(models.Model):
         'State',
         default='draft')
 
-    zone_ids = fields.Many2many(
-        'round.zone',
-        string="Zones",
+    itinerary_ids = fields.Many2many(
+        'round.itinerary',
+        string="Itineraries",
         readonly=True)
 
     picking_ids = fields.One2many(
@@ -85,15 +93,17 @@ class RoundInstance(models.Model):
         )
 
     complete_name = fields.Char(
-        'Display name', readonly=True,
+        'Display Name', readonly=True,
         compute='_get_complete_name', store=True)
 
     @api.multi
-    @api.depends('vehicle_id', 'date', 'time')
+    @api.depends('template_id', 'date', 'time_leave_planned')
     def _get_complete_name(self):
         for rec in self:
             rec.complete_name = '%s %s - %s' % (
-                rec.date, float2time(rec.time), rec.vehicle_id.display_name)
+                rec.date,
+                float2time(rec.time_leave_planned),
+                rec.template_id.display_name)
 
     @api.model
     def create(self, vals):
@@ -103,23 +113,24 @@ class RoundInstance(models.Model):
         return super(RoundInstance, self).create(vals)
 
     @api.multi
-    def button_zone_import(self):
+    def button_itinerary_import(self):
         return dict(self.env.ref(
-            'delivery_rounds.action_round_zone_import').read()[0])
+            'delivery_rounds.action_round_itinerary_import').read()[0])
 
     @api.multi
     def button_update(self):
-        for zone_id in self.zone_ids:
-            self._update_zone(zone_id)
+        for itinerary_id in self.itinerary_ids:
+            self._update_itinerary(itinerary_id)
 
-    def _update_zone(self, zone_id):
-        partner_ids = zone_id.partner_position_ids.mapped('partner_id.id')
+    def _update_itinerary(self, itinerary_id):
+        partner_ids = itinerary_id.partner_position_ids.mapped('partner_id.id')
         # call Try to reserve from stock the qty for confirmed pickings
         picking_confirmed = self.env['stock.picking'].search([
             ('partner_id', 'in', partner_ids),
             ('state', '=', 'confirmed')])
         try:
-            picking_confirmed.action_assign()
+            if not self.env.context.get('skip_reservation'):
+                picking_confirmed.action_assign()
         except UserError:
             # if no moves
             pass
@@ -138,27 +149,22 @@ class RoundInstance(models.Model):
 
     def find(self, partner_id):
         """ Find a delivery_round for this partner """
-        zone_ids = partner_id.round_zone_ids.mapped('zone_id.id')
+        itinerary_ids = partner_id.round_itinerary_ids.mapped(
+            'itinerary_id.id')
         instance = self.search([
             ('state', '=', 'draft'),
-            ('zone_ids', 'in', zone_ids),
+            ('itinerary_ids', 'in', itinerary_ids),
             ], limit=1)
         if instance:
             return instance[0]
-        return None  # do not automatically create new round instance
-        # zone_ids = list(set(zone.id + zone.vehicle_id.zone_ids.ids))
-        # instance = self.create({
-        #     'vehicle': zone.vehicle_id.id,
-        #     'zone_ids': zone_ids
-        #     })  # what date???
-        # return instance
+        return None
 
     count_picking_available_total = fields.Integer(
         'Picking Available Total',
         compute='_get_count_picking',
         readonly=True)
     count_picking_done_total = fields.Integer(
-        'Picking Donee Total',
+        'Picking Done Total',
         compute='_get_count_picking',
         readonly=True)
     count_picking_available_partner = fields.Integer(
@@ -191,6 +197,13 @@ class RoundInstance(models.Model):
             'delivery_rounds.action_picking_tree_available_round').read()[0])
 
     @api.one
+    def button_confirm(self):
+        """ Mark as confirmed. This launch the start of the pickings
+        """
+        self.state = 'open'
+        self.stat_time_picking = time_now(self)
+
+    @api.one
     def button_deliver(self):
         """ Validate all deliveries that are available. Mark as done and unlink
         other deliveries """
@@ -215,6 +228,7 @@ class RoundInstance(models.Model):
         for shipping in self.shipping_ids:
             if shipping.state == 'waiting':
                 shipping.delivery_round_id = False
+        self.stat_time_leave = time_now(self)
 
     @api.multi
     def print_all_deliveryslip(self):