+++ odoo/local-src/delivery_rounds/wizards/instance_itinerary_import.py
@@ -22,11 +22,11 @@
 from odoo import models, fields, api
 
 
-class RoundZoneImport(models.TransientModel):
-    _name = 'round.zone.import'
+class RoundItineraryImport(models.TransientModel):
+    _name = 'round.itinerary.import'
 
-    zone_id = fields.Many2one(
-        'round.zone', 'Zone',
+    itinerary_id = fields.Many2one(
+        'round.itinerary', 'Itinerary',
         required=True)
 
     @api.one
@@ -37,24 +37,26 @@ class RoundZoneImport(models.TransientModel):
             return act_close
         assert len(instance_ids) == 1, "Only 1 ID expected"
         instance = self.env['round.instance'].browse(instance_ids)
-        # partner_ids = self.zone_id.partner_position_ids.mapped(
+        # partner_ids = self.itinerary_id.partner_position_ids.mapped(
         #   'partner_id.id')
         positions = {}
-        for pos in self.zone_id.partner_position_ids:
+        for pos in self.itinerary_id.partner_position_ids:
             positions[pos.partner_id.id] = pos.sequence
 
-        instance.zone_ids += self.zone_id
+        instance.itinerary_ids += self.itinerary_id
 
-        instance._update_zone(self.zone_id)
+        instance._update_itinerary(self.itinerary_id)
 
-        # set sequence on deliveries according to sequence defined in the zone
+        # set sequence on deliveries according to sequence defined in the
+        # itinerary
         shippings = instance.shipping_ids
         # last_seq = max([1] + shippings.mapped('sequence'))
         for shipping in shippings:
             if not shipping.sequence:
                 # shipping.sequence = last_seq + positions[
                 #   shipping.partner_id.id]
-                shipping.sequence = self.zone_id.sequence*10000 + positions[
-                    shipping.partner_id.id]
+                shipping.sequence = (
+                    self.itinerary_id.sequence*10000 +
+                    positions[shipping.partner_id.id])
 
         return act_close